+++ lib/better_chat_scrolling_ai.dart
@@ -1,5 +1,5 @@
 export 'package:scrollable_positioned_list/scrollable_positioned_list.dart'
-    show ItemScrollController, ItemPositionsListener;
+    show ItemScrollController, ItemPositionsListener, ScrollOffsetController;
 
 export 'src/controllers/chat_scroll_controller.dart';
 export 'src/widgets/better_chat_scroll_view.dart';